Actualización de clases y vistas de autenticación
- Cambio de tipo de dato de `Salario` en `Empleado` de `double` a `decimal`.
- Reordenación de propiedades en la clase `Persona`.
- Reemplazo de `DefaultController.cs` por `AuthController.cs` en `Proyecto_CineClub.csproj`.
- Inclusión de `Cine.cs` y `LoginViewModel.cs` en `Proyecto_CineClub.csproj`.
- Inclusión de vistas `Login.cshtml` y `Register.cshtml` en `Proyecto_CineClub.csproj`.
- Añadida cadena de conexión para `CineClub` en `Web.config`.
- Creación del controlador `AuthController` con métodos de autenticación y registro.
- Creación de la clase `LoginViewModel` para el formulario de inicio de sesión.
- Creación de las vistas `Login.cshtml` y `Register.cshtml`.

+++ Proyecto_CineClub/Models/Empleado.cs
@@ -10,7 +10,7 @@ namespace Proyecto_CineClub.Models
         public int IdEmpleado { get; set; }
         public int IdPersona { get; set; }
         public DateTime FechaContratacion { get; set; }        
-        public double Salario { get; set; }
+        public decimal Salario { get; set; }
         public int IdTipoEmpleado { get; set; }        
         public bool EstadoLaboral{ get; set; }
     }
